# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "E-Dashboard API description.",
        "title": "Swagger E-Dashboard API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/edashboard",
    "paths": {
...
        "/inpatient/average-los": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Inpatient"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "vx_user_session_id",
                        "name": "vx_user_session_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "vx_user_name",
                        "name": "vx_user_name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "vx_comp_name",
                        "name": "vx_comp_name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "vx_environment",
                        "name": "vx_environment",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/inpatient/bed-status": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Inpatient"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "vx_user_session_id",
                        "name": "vx_user_session_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "vx_user_name",
                        "name": "vx_user_name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "vx_comp_name",
                        "name": "vx_comp_name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "vx_environment",
                        "name": "vx_environment",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/inpatient/bed-status-popup": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Inpatient"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "vx_user_session_id",
                        "name": "vx_user_session_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "vx_user_name",
                        "name": "vx_user_name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "sBedStatus",
                        "name": "sBedStatus",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "vx_comp_name",
                        "name": "vx_comp_name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "vx_environment",
                        "name": "vx_environment",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/inpatient/demographics": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Inpatient"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "vx_user_session_id",
                        "name": "vx_user_session_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "vx_user_name",
                        "name": "vx_user_name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "sType",
                        "name": "sType",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "vx_comp_name",
                        "name": "vx_comp_name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "vx_environment",
                        "name": "vx_environment",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/inpatient/regadm": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Inpatient"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "vx_user_session_id",
                        "name": "vx_user_session_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "vx_user_name",
                        "name": "vx_user_name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "vx_comp_name",
                        "name": "vx_comp_name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "vx_environment",
                        "name": "vx_environment",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/inpatient/specialty": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Inpatient"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "vx_user_session_id",
                        "name": "vx_user_session_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "vx_user_name",
                        "name": "vx_user_name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "vx_comp_name",
                        "name": "vx_comp_name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "vx_environment",
                        "name": "vx_environment",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/inpatient/ward-census": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Inpatient"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "vx_user_session_id",
                        "name": "vx_user_session_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "vx_user_name",
                        "name": "vx_user_name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "sType",
                        "name": "sType",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "COMPANY_ID",
                        "name": "COMPANY_ID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "vx_environment",
                        "name": "vx_environment",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    }
}